fix: move AddressFormat back to bitcoin module for consistency
- Move AddressFormat from /types back to /utils/blockchain/bitcoin/address.ts
- Follows co-location principle: domain types stay with their domain
- Consistent with AssetInfo staying in counterparty/api.ts
- Update all 34 imports to use bitcoin location
- Remove /types folder as we don't have truly cross-cutting types yet
- AddressFormat is a bitcoin utility type, not a shared application type

This maintains consistency: types belong with their implementation domain.
