+++ src/pages/wallet/import-private-key.tsx
@@ -18,7 +18,7 @@ import { CheckboxInput } from "@/components/inputs/checkbox-input";
 import { PasswordInput } from "@/components/inputs/password-input";
 import { useHeader } from "@/contexts/header-context";
 import { useWallet } from "@/contexts/wallet-context";
-import { AddressFormat } from '@/types';
+import { AddressFormat } from '@/utils/blockchain/bitcoin';
 import { validatePrivateKeyFormat } from "@/utils/validation";
 
 const ImportPrivateKey = () => {
