+++ src/pages/wallet/import-wallet.tsx
@@ -12,7 +12,7 @@ import { CheckboxInput } from "@/components/inputs/checkbox-input";
 import { PasswordInput } from "@/components/inputs/password-input";
 import { useHeader } from "@/contexts/header-context";
 import { useWallet } from "@/contexts/wallet-context";
-import { AddressFormat } from '@/types';
+import { AddressFormat } from '@/utils/blockchain/bitcoin';
 import { isValidCounterwalletMnemonic } from "@/utils/blockchain/counterwallet";
 
 function ImportWallet() {
