+++ src/services/walletService.ts
@@ -1,6 +1,6 @@
 import { defineProxyService } from '@webext-core/proxy-service';
 import { sendMessage } from 'webext-bridge/background'; // Import for background context
-import { AddressFormat } from '@/types';
+import { AddressFormat } from '@/utils/blockchain/bitcoin';
 import { walletManager, settingsManager, type Wallet, type Address } from '@/utils/wallet';
 
 interface WalletService {
