--- src/types/index.ts
@@ -1,6 +1,0 @@
-/**
- * Central types export for the application.
- * These are shared types used across multiple domains.
- */
-
-export * from './bitcoin';
