+++ src/utils/blockchain/bitcoin/__tests__/privateKey.test.ts
@@ -7,7 +7,7 @@ import {
   getAddressFromPrivateKey,
   getPrivateKeyFromMnemonic
 } from '@/utils/blockchain/bitcoin/privateKey';
-import { AddressFormat } from '@/types';
+import { AddressFormat } from '@/utils/blockchain/bitcoin';
 
 vi.mock('@/utils/blockchain/counterwallet', () => ({
   getCounterwalletSeed: vi.fn(() => new Uint8Array(64).fill(1))
