+++ src/utils/blockchain/bitcoin/__tests__/transactionSigner.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { signTransaction } from '@/utils/blockchain/bitcoin/transactionSigner';
-import { AddressFormat } from '@/types';
+import { AddressFormat } from '@/utils/blockchain/bitcoin';
 import { Transaction } from '@scure/btc-signer';
 import { hexToBytes, bytesToHex } from '@noble/hashes/utils';
 import type { Wallet, Address } from '@/utils/wallet';
