+++ src/utils/blockchain/bitcoin/privateKey.ts
@@ -6,7 +6,7 @@ import { HDKey } from '@scure/bip32';
 import { generateMnemonic, mnemonicToSeedSync } from '@scure/bip39';
 import { wordlist } from '@scure/bip39/wordlists/english';
 import { encodeAddress } from '@/utils/blockchain/bitcoin/address';
-import { AddressFormat } from '@/types';
+import { AddressFormat } from '@/utils/blockchain/bitcoin';
 import { getCounterwalletSeed } from '@/utils/blockchain/counterwallet';
 
 // Create a base58check encoder instance for WIF usage.
