+++ src/utils/encryption/walletEncryption.ts
@@ -1,6 +1,6 @@
 import { validateMnemonic } from '@scure/bip39';
 import { wordlist } from '@scure/bip39/wordlists/english';
-import { AddressFormat } from '@/types';
+import { AddressFormat } from '@/utils/blockchain/bitcoin';
 import { isValidCounterwalletMnemonic } from '@/utils/blockchain/counterwallet';
 import { encryptString, decryptString, DecryptionError } from '@/utils/encryption/encryption';
