+++ src/utils/storage/__tests__/walletStorage.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, vi, type Mock } from 'vitest';
-import { AddressFormat } from '@/types';
+import { AddressFormat } from '@/utils/blockchain/bitcoin';
 import {
   getAllRecords,
   addRecord,
