+++ src/utils/validation/__tests__/privateKey.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import * as fc from 'fast-check';
-import { AddressFormat } from '@/types';
+import { AddressFormat } from '@/utils/blockchain/bitcoin';
 import {
   validatePrivateKeyFormat,
   sanitizePrivateKey,
